extract common functions
